argo/kfp-compiler/pyproject.toml
[tool.poetry]
name = "kfp_compiler"
version = "0.0.0" # set by poetry-dynamic-versioning
description = ""
authors = []

[tool.poetry.dependencies]
python = "3.9.13"
click = ">=7,<8"
PyYAML = "^5.4.1"
kfp = "1.8.13"
libclang = ">=13.0.0"
apache-beam = "2.40.0"
tfx = ">=1.9.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"


[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"
format-jinja = "{{base}}+{{distance}}.{{commit}}{% if dirty %}.dirty{% endif %}"

[build-system]
requires = ["poetry-core>=1.1.9", "poetry-dynamic-versioning"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
kfp-compiler = 'kfp_compiler.__main__:main'
